Fix new search page window issues: - crash - background color - window insets
