+++ app/shared/ui-exploration/src/androidMain/kotlin/ui/exploration/search/SearchPage.android.kt
@@ -100,7 +100,7 @@ fun PreviewSearchPageResultColumn() = ProvideCompositionLocalsForPreview {
             selectedItemIndex = { 1 },
             onSelect = {},
             onPlay = {},
-            searchBar = {},
+            headers = {},
         )
     }
 }
